added delete transaction logic

+++ src/services/DeleteTransactionService.ts
@@ -1,11 +1,17 @@
-// import AppError from '../errors/AppError';
-
 import { getRepository } from 'typeorm';
+import AppError from '../errors/AppError';
+
 import Transaction from '../models/Transaction';
 
 class DeleteTransactionService {
   public async execute(id: string): Promise<void> {
     const transactionRepository = getRepository(Transaction);
+
+    const transaction = await transactionRepository.findOne(id);
+
+    if (!transaction) {
+      throw new AppError('Transaction does not exist');
+    }
 
     await transactionRepository.delete({ id });
   }
